Don't update document stats for non-idle views
This functionality should be merged into the DocumentTimerManager,
which itself should run the different document idle tasks via
seperate jobs instead of a single idle, if they don't depend on
each other.

As a workaround, this checks for the idle status of the view and
won't run on busy documents.

The idle is also changed to a timer, which sleeps for busy views.

Change-Id: I185137ed3423ecaae0f7edb39018d26c4244d359